Fixed bug on craft loading.
SDL2 updated to version 2.0.2.

Fixed OSX rendering.
Allow vertex arrays on OSX.

Refactoring.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@593 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 8be49b3f8d98f72173c9aa4fd28c5637328b3e5c

+++ Source/SpriteManagerFont.cpp
@@ -149,7 +149,7 @@ void SpriteManager::BuildFonts()
 {
     for( size_t i = 0; i < Fonts.size(); i++ )
         if( Fonts[ i ] && !Fonts[ i ]->Builded )
-            BuildFont( i );
+            BuildFont( (int) i );
 }
 
 void SpriteManager::BuildFont( int index )
